add test fixtures to root conftest as pytest_plugins

black & isort
