+++ tests/conftest.py
@@ -2,6 +2,8 @@ from typing import Dict, Tuple
 
 import pytest
 
+pytest_plugins = "tests.fixtures"
+
 
 @pytest.fixture(scope="session")
 def monkeysession(request):
